Move accounting methods to Account
