+++ nautilus_trader/model/instrument.pyx
@@ -21,9 +21,6 @@ from decimal import Decimal
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.model.c_enums.asset_class cimport AssetClass
 from nautilus_trader.model.c_enums.asset_type cimport AssetType
-from nautilus_trader.model.c_enums.liquidity_side cimport LiquiditySide
-from nautilus_trader.model.c_enums.liquidity_side cimport LiquiditySideParser
-from nautilus_trader.model.c_enums.position_side cimport PositionSide
 from nautilus_trader.model.currency cimport Currency
 from nautilus_trader.model.identifiers cimport InstrumentId
 from nautilus_trader.model.identifiers cimport Symbol
@@ -241,198 +238,6 @@ cdef class Instrument(Data):
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}('{self.id.value}')"
-
-    cpdef Money market_value(
-        self,
-        Quantity quantity,
-        close_price: Decimal,
-        leverage: Decimal=None,
-    ):
-        """
-        Calculate the market value from the given parameters.
-
-        Parameters
-        ----------
-        quantity : Quantity
-            The total quantity.
-        close_price : Decimal or Price
-            The closing price.
-        leverage : Decimal, optional
-            The leverage for the position.
-
-        Returns
-        -------
-        Money
-            In the settlement currency.
-
-        """
-        if leverage is None:
-            leverage = Decimal(1)
-        Condition.not_none(quantity, "quantity")
-        Condition.type(close_price, (Decimal, Price), "close_price")
-        Condition.not_none(close_price, "close_price")
-        Condition.not_none(leverage, "leverage")
-
-        if self.is_inverse:
-            close_price = 1 / close_price
-
-        market_value: Decimal = (quantity * close_price * self.multiplier) / leverage
-        return Money(market_value, self.settlement_currency)
-
-    cpdef Money notional_value(self, Quantity quantity, close_price: Decimal):
-        """
-        Calculate the notional value from the given parameters.
-
-        Parameters
-        ----------
-        quantity : Quantity
-            The total quantity.
-        close_price : Decimal or Price
-            The closing price.
-
-        Returns
-        -------
-        Money
-            In the settlement currency.
-
-        """
-        Condition.not_none(quantity, "quantity")
-        Condition.type(close_price, (Decimal, Price), "close_price")
-        Condition.not_none(close_price, "close_price")
-
-        if self.is_inverse:
-            close_price = 1 / close_price
-
-        notional_value: Decimal = quantity * close_price * self.multiplier
-        return Money(notional_value, self.settlement_currency)
-
-    cpdef Money calculate_initial_margin(
-        self,
-        Quantity quantity,
-        Price price,
-        leverage: Decimal=None,
-    ):
-        """
-        Calculate the initial margin from the given parameters.
-
-        Parameters
-        ----------
-        quantity : Quantity
-            The order quantity.
-        price : Price
-            The order price.
-        leverage : Decimal, optional
-            The leverage for the position.
-
-        Returns
-        -------
-        Money
-            In the settlement currency.
-
-        """
-        if leverage is None:
-            leverage = Decimal(1)
-        Condition.not_none(quantity, "quantity")
-        Condition.not_none(price, "price")
-        Condition.not_none(leverage, "leverage")
-
-        if leverage == 1:
-            return Money(0, self.settlement_currency)  # No margin necessary
-
-        notional = self.notional_value(quantity, price)
-        margin = notional / leverage * self.margin_init
-        margin += notional * self.taker_fee * 2
-
-        return Money(margin, self.settlement_currency)
-
-    cpdef Money calculate_maint_margin(
-        self,
-        PositionSide side,
-        Quantity quantity,
-        Price last,
-        leverage: Decimal=None,
-    ):
-        """
-        Calculate the maintenance margin from the given parameters.
-
-        Parameters
-        ----------
-        side : PositionSide
-            The currency position side.
-        quantity : Quantity
-            The currency position quantity.
-        last : Price
-            The position instruments last price.
-        leverage : Decimal, optional
-            The leverage for the position.
-
-        Returns
-        -------
-        Money
-            In the settlement currency.
-
-        """
-        if leverage is None:
-            leverage = Decimal(1)
-        # side checked in _get_close_price
-        Condition.not_none(quantity, "quantity")
-        Condition.not_none(last, "last")
-        Condition.not_none(leverage, "leverage")
-
-        if leverage == 1:
-            return Money(0, self.settlement_currency)  # No margin necessary
-
-        notional = self.notional_value(quantity, last)
-        margin = (notional / leverage) * self.margin_maint
-        margin += notional * self.taker_fee
-
-        return Money(margin, self.settlement_currency)
-
-    cpdef Money calculate_commission(
-        self,
-        Quantity last_qty,
-        last_px: Decimal,
-        LiquiditySide liquidity_side,
-    ):
-        """
-        Calculate the commission generated from a transaction with the given
-        parameters.
-
-        Parameters
-        ----------
-        last_qty : Quantity
-            The transaction quantity.
-        last_px : Decimal or Price
-            The transaction price.
-        liquidity_side : LiquiditySide
-            The liquidity side for the transaction.
-
-        Returns
-        -------
-        Money
-            In the settlement currency.
-
-        Raises
-        ------
-        ValueError
-            If liquidity_side is NONE.
-
-        """
-        Condition.not_none(last_qty, "last_qty")
-        Condition.type(last_px, (Decimal, Price), "last_px")
-        Condition.not_equal(liquidity_side, LiquiditySide.NONE, "liquidity_side", "NONE")
-
-        notional: Decimal = self.notional_value(quantity=last_qty, close_price=last_px)
-
-        if liquidity_side == LiquiditySide.MAKER:
-            commission: Decimal = notional * self.maker_fee
-        elif liquidity_side == LiquiditySide.TAKER:
-            commission: Decimal = notional * self.taker_fee
-        else:
-            raise RuntimeError(f"invalid LiquiditySide, "
-                               f"was {LiquiditySideParser.to_str(liquidity_side)}")
-
-        return Money(commission, self.settlement_currency)
 
 
 # # TODO: Finish docs
