+++ tests/unit_tests/trading/test_trading_account.py
@@ -13,6 +13,7 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+from decimal import Decimal
 import unittest
 
 from nautilus_trader.common.clock import TestClock
@@ -22,14 +23,27 @@ from nautilus_trader.core.uuid import uuid4
 from nautilus_trader.data.cache import DataCache
 from nautilus_trader.model.currencies import BTC
 from nautilus_trader.model.currencies import ETH
+from nautilus_trader.model.currencies import JPY
 from nautilus_trader.model.currencies import USD
+from nautilus_trader.model.currencies import USDT
+from nautilus_trader.model.enums import LiquiditySide
+from nautilus_trader.model.enums import PositionSide
 from nautilus_trader.model.events import AccountState
 from nautilus_trader.model.identifiers import AccountId
 from nautilus_trader.model.identifiers import StrategyId
 from nautilus_trader.model.identifiers import TraderId
+from nautilus_trader.model.identifiers import Venue
 from nautilus_trader.model.objects import Money
+from nautilus_trader.model.objects import Price
+from nautilus_trader.model.objects import Quantity
 from nautilus_trader.trading.account import Account
 from nautilus_trader.trading.portfolio import Portfolio
+from tests.test_kit.providers import TestInstrumentProvider
+
+
+AUDUSD_SIM = TestInstrumentProvider.default_fx_ccy("AUD/USD")
+USDJPY_SIM = TestInstrumentProvider.default_fx_ccy("USD/JPY")
+BTCUSDT_BINANCE = TestInstrumentProvider.btcusdt_binance()
 
 
 class AccountTests(unittest.TestCase):
@@ -45,6 +59,122 @@ class AccountTests(unittest.TestCase):
 
         self.portfolio = Portfolio(self.clock, logger)
         self.portfolio.register_cache(DataCache(logger))
+
+    def test_calculate_order_margin_with_no_leverage_returns_zero(self):
+        # Arrange
+        instrument = TestInstrumentProvider.xbtusd_bitmex()
+
+        margin = Account.calculate_initial_margin(
+            instrument,
+            Quantity(100000),
+            Price("11493.60"),
+        )
+
+        # Assert
+        self.assertEqual(Money("0.00000000", BTC), margin)
+
+    def test_calculate_position_margin_with_no_leverage_returns_zero(self):
+        # Arrange
+        instrument = TestInstrumentProvider.xbtusd_bitmex()
+
+        # Act
+        margin = Account.calculate_maint_margin(
+            instrument,
+            PositionSide.LONG,
+            Quantity(100000),
+            Price("11493.60"),
+        )
+
+        # Assert
+        self.assertEqual(Money("0.00000000", BTC), margin)
+
+    def test_calculate_notional_value(self):
+        # Arrange
+        instrument = TestInstrumentProvider.btcusdt_binance()
+
+        # Act
+        value = Account.notional_value(
+            instrument,
+            Quantity(10),
+            Price("11493.60"),
+        )
+
+        # Assert
+        self.assertEqual(Money("114936.00000000", USDT), value)
+
+    def test_calculate_notional_value_for_inverse(self):
+        # Arrange
+        instrument = TestInstrumentProvider.xbtusd_bitmex()
+
+        # Act
+        value = Account.notional_value(
+            instrument,
+            Quantity(100000),
+            Price("11493.60"),
+        )
+
+        # Assert
+        self.assertEqual(Money("8.70049419", BTC), value)
+
+    def test_calculate_commission_for_maker_crypto(self):
+        # Arrange
+        instrument = TestInstrumentProvider.xbtusd_bitmex()
+
+        # Act
+        commission = Account.calculate_commission(
+            instrument,
+            Quantity(100000),
+            Decimal("11450.50"),
+            LiquiditySide.MAKER,
+        )
+
+        # Assert
+        self.assertEqual(Money("-0.00218331", BTC), commission)
+
+    def test_calculate_commission_for_taker_fx(self):
+        # Arrange
+        instrument = TestInstrumentProvider.default_fx_ccy("AUD/USD", Venue("IDEALPRO"))
+
+        # Act
+        commission = Account.calculate_commission(
+            instrument,
+            Quantity(1500000),
+            Decimal("0.80050"),
+            LiquiditySide.TAKER,
+        )
+
+        # Assert
+        self.assertEqual(Money(24.02, USD), commission)
+
+    def test_calculate_commission_crypto_taker(self):
+        # Arrange
+        instrument = TestInstrumentProvider.xbtusd_bitmex()
+
+        # Act
+        commission = Account.calculate_commission(
+            instrument,
+            Quantity(100000),
+            Decimal("11450.50"),
+            LiquiditySide.TAKER,
+        )
+
+        # Assert
+        self.assertEqual(Money("0.00654993", BTC), commission)
+
+    def test_calculate_commission_fx_taker(self):
+        # Arrange
+        instrument = TestInstrumentProvider.default_fx_ccy("USD/JPY", Venue("IDEALPRO"))
+
+        # Act
+        commission = Account.calculate_commission(
+            instrument,
+            Quantity(2200000),
+            Decimal("120.310"),
+            LiquiditySide.TAKER,
+        )
+
+        # Assert
+        self.assertEqual(Money(5293.64, JPY), commission)
 
     def test_instantiated_accounts_basic_properties(self):
         # Arrange
